Implement base steps for game end
Also handle the cases in which one of the two players leaves
the game before the end.

todo:
	- Create the match when a game ends and check who is the
	winner.
	- Add the timer

NB. the problem on lobby page reload still remain (when the
player is listening for the join). In this case only the
joiner passes to the game.
I tried a solution but it does not actually work.